+++ tests/test_autopost.py
@@ -21,16 +21,12 @@ def autopost(session: ClientSession) -> AutoPoster:
 
 
 @pytest.mark.asyncio
-async def test_AutoPoster_breaks_autopost_loop_on_401(
-    mocker: MockerFixture, session: ClientSession
-) -> None:
+async def test_AutoPoster_breaks_autopost_loop_on_401(mocker: MockerFixture, session: ClientSession) -> None:
     response = mock.Mock("reason, status")
     response.reason = "Unauthorized"
     response.status = 401
 
-    mocker.patch(
-        "topgg.DBLClient.post_guild_count", side_effect=Unauthorized(response, {})
-    )
+    mocker.patch("topgg.DBLClient.post_guild_count", side_effect=Unauthorized(response, {}))
 
     callback = mock.Mock()
     autopost = DBLClient("", session=session).autopost().stats(callback)
@@ -46,9 +42,7 @@ async def test_AutoPoster_breaks_autopost_loop_on_401(
 
 @pytest.mark.asyncio
 async def test_AutoPoster_raises_missing_stats(autopost: AutoPoster) -> None:
-    with pytest.raises(
-        TopGGException, match="you must provide a callback that returns the stats."
-    ):
+    with pytest.raises(TopGGException, match="you must provide a callback that returns the stats."):
         await autopost.start()
 
 
@@ -66,9 +60,7 @@ async def test_AutoPoster_interval_too_short(autopost: AutoPoster) -> None:
 
 
 @pytest.mark.asyncio
-async def test_AutoPoster_error_callback(
-    mocker: MockerFixture, autopost: AutoPoster
-) -> None:
+async def test_AutoPoster_error_callback(mocker: MockerFixture, autopost: AutoPoster) -> None:
     error_callback = mock.Mock()
     response = mock.Mock("reason, status")
     response.reason = "Internal Server Error"