+++ docs/whats_new.rst
@@ -8,6 +8,18 @@ What's New
 
 This page keeps a detailed human friendly rendering of what's new and changed in specific versions.
 
+v2.0.1
+======
+* Added Python 3.12 support
+* Dropped Python 3.6 and 3.7 support
+* Removed the need to manually set a ``default_bot_id`` property
+* :attr:`BotData.def_avatar` is now an optional string
+* :meth:`DBLClient.get_bots` is now deprecated
+* :meth:`DBLClient.get_guild_count` no longer accepts a ``bot_id`` argument
+* :meth:`DBLClient.get_bot_votes` no longer raises a ``ClientException`` without a ``default_bot_id`` property
+* :meth:`DBLClient.get_bot_info` no longer raises a ``ClientException`` without a ``default_bot_id`` property
+* :meth:`DBLClient.generate_widget`` no longer raises a ``ClientException`` without a ``default_bot_id`` property
+
 v2.0.0a
 =======
 * :obj:`~.DBLClient` now doesn't take in ``discord.Client`` instance